+++ tests/__init__.py
@@ -1,1 +1,2 @@
-from .test_all import *
+from .test_create import *
+from .test_parse import *